% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covr_adapted.R
\name{adapted_check}
\alias{adapted_check}
\title{Use devtools::check running testthat::skip_on_cran() tests}
\usage{
adapted_check(..., run_tests = "true")
}
\arguments{
\item{...}{arguments passed to devtools::check}

\item{run_tests}{value of the NOT_CRAN environment variable. This should be
either 'true' or 'false'}
}
\value{
results from devtools::check
}
\description{
Use devtools::check running testthat::skip_on_cran() tests
}
\author{
Jorge Cimentada
}
