% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covr_adapted.R
\name{adapted_covr}
\alias{adapted_covr}
\title{Use covr::codecov while running testthat::skip_on_cran() tests}
\usage{
adapted_covr(..., run_tests = "true")
}
\arguments{
\item{...}{arguments passed to covr::codecov}

\item{run_tests}{value of the NOT_CRAN environment variable. This should be
either 'true' or 'false'}
}
\value{
results from covr::codecov
}
\description{
Use covr::codecov while running testthat::skip_on_cran() tests
}
\author{
Jorge Cimentada
}
