% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_covr_adapted.R
\name{adapted_test}
\alias{adapted_test}
\title{Use devtools::test running testthat::skip_on_cran() tests}
\usage{
adapted_test(..., run_tests = "true")
}
\arguments{
\item{...}{arguments passed to devtools::test}

\item{run_tests}{value of the NOT_CRAN environment variable. This should be
either 'true' or 'false'}
}
\value{
results from devtools::test
}
\description{
Use devtools::test running testthat::skip_on_cran() tests
}
\author{
Jorge Cimentada
}
