% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_citation.R
\name{add_citation}
\alias{add_citation}
\title{Add all citation information of your package}
\usage{
add_citation()
}
\value{
Nothing
}
\description{
Adds the citation skeleton in \code{inst/CITATION} filled out
automatically with the information from the \code{DESCRIPTION} file.
Additionally, it creates an \code{.onAttach} function on 'R/utils.R'
(created if necessary, appended if exists) which prints out the citation
details on startup.
}
\author{
Jorge
}
