% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proj.R
\name{create_proj}
\alias{create_proj}
\title{Create  opinionated package structure using `drake` and `renv`}
\usage{
create_proj(repo_name, path = ".")
}
\arguments{
\item{repo_name}{Name of the repository in github. For example
"cimentadaj/fragile_families".}

\item{path}{the path where the repository will be saved locally}
}
\value{
invisible TRUE
}
\description{
Create an opinionated package structure using `drake` and `renv`
}
\details{
Clones `repo_name` from your Github account and saves it in the
`path` specified in the argument. Once cloned, the function
creates files `R/plan.R`, `R/functions.R`, `R/plan.R` and `_drake.R`
as per the `drake` project structure. It adds a few files to ignore
 in .gitignore, creates and `Rmd` for the README and adds a template
of code for `_drake.R` and `R/plan.R`.

Finally, it ends by installing `renv` for package dependency, installs
`drake` automatically and restarts the session in the project directory
to being working.
}
\examples{

\dontrun{
create_proj("cimentadaj/fragile_families", ".")
}

}
\author{
Jorge
}
