% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.data.frame}
\alias{describe.data.frame}
\title{Describe method for data frames}
\usage{
\method{describe}{data.frame}(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A tibble with usual properties such as length, size, and attributes such as names, dimensions
if apply.
}
\description{
Describe method for data frames
}
