% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.default}
\alias{describe.default}
\title{Describe method for unknown object classes}
\usage{
\method{describe}{default}(x)
}
\arguments{
\item{x}{An object}
}
\value{
Text specifying that no method exists for that class
}
\description{
Describe method for unknown object classes
}
