% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.generic}
\alias{describe.generic}
\title{Describe method for atomic vectors}
\usage{
\method{describe}{generic}(x)
}
\arguments{
\item{x}{A numeric, integer, double or character vector}
}
\value{
A tibble with usual properties such as length, size, and attributes such as names, dimensions
if apply.
}
\description{
Describe method for atomic vectors
}
