% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.list}
\alias{describe.list}
\title{Describe method for lists}
\usage{
\method{describe}{list}(x)
}
\arguments{
\item{x}{A list}
}
\value{
A tibble with usual properties such as length, size, and attributes such as names, dimensions
if apply.
}
\description{
Describe method for lists
}
