% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.matrix}
\alias{describe.matrix}
\title{Describe method for matrices}
\usage{
\method{describe}{matrix}(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A tibble with usual properties such as length, size, and attributes such as names, dimensions
if apply.
}
\description{
Describe method for matrices
}
