% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Get a description of object x}
\usage{
describe(x)
}
\arguments{
\item{x}{Any object. Currently supported are all R classes, plus matrices, data frames and lists.}
}
\value{
A tibble with usual properties such as length, size, and attributes such as names, dimensions
if apply
}
\description{
Get a description of object x
}
\examples{


\dontrun{
x <- NULL
describe(x)

x <- numeric()
describe(x)

x <- "hey"
describe(x)

x <- seq(0.01, 0.05, 0.01)
describe(x)

x <- 1:10
describe(x)

x <- matrix()
describe(x)

x <- mtcars
describe(x)

x <- data.frame()
describe(x)

x <- list(list(1:10), 1:2)
describe(x)

x <- list()
describe(x)

x <- list(1:10, 1:2)
describe(x)


}
}
