% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rscript.R
\name{gen_rscript}
\alias{gen_rscript}
\title{Generate an R script with a preamble}
\usage{
gen_rscript(dir, filename, author, project, open_file = F)
}
\arguments{
\item{dir}{The directory where the new script will be stored as a character string}

\item{filename}{The R script file name as a character string (without extensions)}

\item{author}{The author of the script as a character}

\item{project}{Project name or script name that will appear in the preamble as character string.}

\item{open_file}{Logical to whether you want to open the file with Rstudio after the function call}
}
\value{
Returns nothing. If \code{open_file} is \code{TRUE}, it will open the script.
}
\description{
Generate an R script with a preamble
}
\examples{
# Only been tested with OSx El Capitan 10.11.6

gen_rscript(dir = ".", "new_script", "Jorge Cimentada", "new_project", FALSE)
# This should create a new R script in your working directory

}

