% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Build sequence of replicate-survey-adjusted models}
\usage{
models(dv, covariates, data)
}
\arguments{
\item{dv}{A character element specifying the name of the dependent variable}

\item{covariates}{A character vector with the names of the covariates}

\item{data}{The replicated survey design which contains all of the variables outlined above}
}
\value{
Returns a list with length(covariates) where the first slot are the results of the dependent variable
regressed on the first covariate. The second slot is the dependent variable regressed on the first two
 covariates and so on, until the last slot, which contains the dependent variable regressed on all covariates
}
\description{
Build sequence of replicate-survey-adjusted models
}
\examples{
# Don't have time
}

