% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_new_post.R
\name{my_new_post}
\alias{my_new_post}
\title{Insert a new post using blogdown}
\usage{
my_new_post(title, ...)
}
\arguments{
\item{title}{Your blogpost's title}

\item{...}{All extra arguments passed to blogdown::new_post}
}
\value{
It creates a .Rmd and .html file under ./blog/your-title-name/
}
\description{
Insert a new post using blogdown
}
\examples{

# No time

}
