% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa_countrynames.R
\docType{data}
\name{pisa_countrynames}
\alias{pisa_countrynames}
\title{PISA search table for all country names}
\usage{
pisa_countrynames
}
\description{
This vector contains all abbreviated country names for all PISA
waves (2000 until 2015) with their correctly written equivalent.
Some of these abbreviations are in ISO 3-letter while others
were written incorrectly, e.g. Korea, Republic of.
}
\examples{
# install.packages(PISA2006lite)
library(PISA2006lite)

unique(PISA2006lite::student2006$CNT)
unique(pisa_countrynames[as.character(PISA2006lite::student2006$CNT)]) # New country names
# This can be saved as a new column or replace the previous column.
}
\keyword{pisa_countrynames}

