% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_name.R
\name{reverse_name}
\alias{reverse_name}
\title{reverse_name}
\usage{
reverse_name(x)
}
\arguments{
\item{x}{a named vector}
}
\value{
same vector with the vector names as elements and elements as names
}
\description{
reverse_name
}
\examples{
reverse_name(c(a = 'b', c = 'b'))
}
