% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_se.R
\name{robustse}
\alias{robustse}
\title{Estimate robust standard errors for different types of coefficients}
\usage{
robustse(x, coef = c("logit", "odd.ratio", "probs"))
}
\arguments{
\item{x}{A linear or glm models}

\item{coef}{A character vector indicating the type of coefficient. The function accepts 'logit' for
log coefficients, 'odd.ratio' for odd ratios and 'probs' for probability scaled coefficients.}
}
\value{
A matrix containing the estimated coefficient, robust standard errors, z values and p values
}
\description{
Estimate robust standard errors for different types of coefficients
}
\examples{
library(haven)
dat <- read_dta("http://www.stata-press.com/data/r9/quad1.dta")
mod1 <- glm(z ~ x1 + x2 + x3, dat, family = binomial)

robustse(mod1, coef = "logit")
robustse(mod1, coef = "odd.ratio")
robustse(mod1, coef = "probs")

# See http://www.jorgecimentada.com/?p=214 for a detailed example.
}
