% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer2.R
\name{stargazer2}
\alias{stargazer2}
\title{Logistic regression table with odd ratios}
\usage{
stargazer2(model, odd.ratio = FALSE, ...)
}
\arguments{
\item{model}{A non-empty list containing one or more glm models}

\item{odd.ratio}{A logical value indicating whether you want odd ratios (TRUE)
instead of the usual log odds (FALSE).}

\item{...}{further aguments passed to the stargazer function. See ?stargazer::stargazer}
}
\value{
A stargazer table with correct odd ratios, standard errors and p values
}
\description{
Logistic regression table with odd ratios
}
\examples{
stargazer2(glm(am ~ mpg, data = mtcars), odd.ratio = TRUE, type = "text") # Odd ratios
stargazer2(glm(am ~ mpg, data = mtcars), odd.ratio = FALSE, type = "text") # Log odds

# Let's compare it with Stata output:

# R table
library(haven)
auto <- haven::read_dta("http://www.stata-press.com/data/r13/auto.dta")
auto$mpg2 <- auto$mpg > 21.29
model <- glm(mpg2 ~ price + rep78 + turn + displacement, data = auto, family = "binomial")
stargazer2(model, odd.ratio = TRUE, type = "text", single.row = TRUE)

# Stata table (run in Stata)
# use http://www.stata-press.com/data/r13/auto.dta, clear
# gen mpg2 = mpg > 21.29
# logit mpg2 price rep78 turn displacement, or

}

