% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{A stopwatch}
\usage{
timer(mins, sound = 8)
}
\arguments{
\item{mins}{a number. can be minutes or seconds.}

\item{sound}{a number specifying the ring sound. See ?beepr::beep}
}
\value{
A ringtone will go off when the time is off.
}
\description{
A stopwatch
}
\examples{
# Not run
# Not run
}
